Moved find and calibrate methods from CCL qubit to qubit_object
