+++ pycqed/instrument_drivers/meta_instrument/qubit_objects/qubit_object.py
@@ -258,6 +258,324 @@ class Qubit(Instrument):
                        MC=None, analyze=True, close_fig=True):
         raise NotImplementedError()
 
+    def find_resonators(self, start_freq=7e9, stop_freq=8e9, VNA_power=-40,
+                        bandwidth=200, timeout=200, npts=2001, with_VNA=None,
+                        verbose=True):
+        """
+        Performs a wide range scan to find all resonator dips. Will use VNA if one
+        is connected and linked to the qubit object, or if specified via 'with_VNA'.
+        
+        TODO: Add measure_with_VNA to CCL Transmon object
+        """
+        if with_VNA is None:
+            try:
+                if self.instr_VNA.get_instr() == '':
+                    with_VNA = False
+                else:
+                    with_VNA = True
+            except:
+                with_VNA = False
+
+        if with_VNA:
+            VNA = self.instr_VNA.get_instr()
+
+            VNA.start_frequency(start_freq)
+            VNA.stop_frequency(stop_freq)
+            VNA.power(VNA_power)
+            VNA.bandwidth(bandwidth)
+            VNA.npts(npts)
+            # VNA.timeout(timeout)
+            name = 'Initial_VNA'
+
+            self.measure_with_VNA(VNA, name=name)
+
+            result = ma2.sa.Initial_Resonator_Scan_Analysis(label=name)
+        else:
+            self.ro_pulse_amp_CW(0.01)
+            freqs = np.linspace(start_freq, stop_freq, npts)
+            self.measure_heterodyne_spectroscopy(freqs=freqs, analyze=False)
+            result = ma2.sa.Initial_Resonator_Scan_Analysis()
+
+        peak_freqs = []
+        for peak in result.peaks:
+            if peak not in peak_freqs:
+                peak_freqs.append(peak)
+
+        self.res_dict = {}
+        for i, freq in enumerate(result.peaks):
+            self.res_dict[str(i)] = [freq, 'unknown', {}, None, 0]
+
+        if verbose:
+            for resonator, items in self.res_dict.items():
+                print('{}:\t{:.3f} GHz'.format(resonator, items[0]/1e9))
+
+        return True
+
+    def calibrate_spec_pow(self, freqs=None, start_power=-35, threshold=0.1,
+                           verbose=True):
+        """
+        Finds the optimal spectroscopy power for qubit spectroscopy (not pulsed)
+        by varying it in steps of 5 dBm, and ending when the peak has power 
+        broadened by 1+threshold (default: broadening of 10%)
+        """
+        if freqs is None:
+            freqs = np.arange(self.freq_qubit() - 50e6,
+                              self.freq_qubit() + 50e6, 1e6)
+        power = start_power
+
+        w0, w = 1e9, 1e9
+
+        while w < (1 + threshold) * w0:
+            self.spec_pow(power)
+            self.measure_spectroscopy(freqs=freqs, analyze=False,
+                                      label='spec_pow_' + str(power) + '_dBm')
+
+            a = ma.Qubit_Spectroscopy_Analysis(label=self.msmt_suffix,
+                                               qb_name=self.name)
+
+            w = a.params['kappa'].value
+            power += 5
+            if w < w0:
+                w0 = w
+        if verbose:
+            print('setting spectroscopy power to {}'.format(power-5))
+        self.spec_pow(power-5)
+        return True
+
+    def find_resonator_frequency_initial(self, start_freq=7e9, stop_freq=8e9,
+                                         npts=50001, use_min=False, MC=None,
+                                         update=True, with_VNA=None):
+        '''
+        quick script that uses measure_heterodyne_spectroscopy on a wide range
+        to act as a sort of mock of a VNA resonator scan'.
+
+        If it is a first scan (no freq_res yet in qubit object) it will perform
+        a wide range scan. Otherwise it will zoom in on a resonator
+        '''
+        if with_VNA is None:
+            try:
+                if self.instr_VNA.get_instr() == '':
+                    with_VNA = False
+                else:
+                    with_VNA = True
+            except:
+                with_VNA = False
+
+        delkeys = []
+        for resonator, items in self.res_dict.items():
+            freq = items[0]
+            if with_VNA:
+                VNA = self.instr_VNA.get_instr()
+                start_freq = freq - 10e6
+                stop_freq = freq + 10e6
+                name = 'VNA_Resonator_scan_' + str(round(freq/1e9, 4)) + 'GHz'
+                self.measure_with_VNA(VNA, start_freq, stop_freq, npts)
+            else:
+                freqs = np.arange(freq-5e6, freq+5e6, 0.1e6)
+                name = 'Resonator_scan' + self.msmt_suffix
+                self.measure_heterodyne_spectroscopy(freqs=freqs,
+                                                     analyze=False)
+
+            a = ma.Homodyne_Analysis(label=name, qb_name=self.name)
+
+            dip = np.amin(a.data_y)
+            offset = a.fit_results.params['A'].value
+
+            if np.abs(dip/offset) > 0.5:
+                print('Removed candidate {} ({:.3f} GHz): Not a resonator'
+                      .format(resonator, freq/1e9))
+                delkeys.append(resonator)
+            elif np.isnan(a.fit_results.params['Qc'].stderr):
+                print('Removed candidate {} ({:.3f} GHz): Not a resonator'
+                      .format(resonator, freq/1e9))
+                delkeys.append(resonator)
+            else:
+                if use_min:
+                    f_res = a.min_frequency
+                else:
+                    f_res = a.fit_results.params['f0'].value*1e9
+
+                # Check if not a duplicate
+                i = int(resonator)
+                if i > 0:
+                    prev_freq = self.res_dict[str(i-1)][0]
+                    if np.abs(prev_freq - f_res) < 10e6:
+                        delkeys.append(resonator)
+                        print('Removed candidate: ' + resonator + ' (' +
+                              str(round(f_res/1e9, 3)) + ' GHz): Duplicate')
+                    else:
+                        self.res_dict[resonator][0] = f_res
+                        print("Added resonator " + resonator + ' (' +
+                              str(round(f_res/1e9, 3)) + ' GHz)')
+
+        for delkey in delkeys:
+            self.res_dict.pop(delkey)
+
+        # Rearrange dictionary to start from 0 again:
+        i = 0
+        newdict = {}
+        for resonator, items in self.res_dict.items():
+            newdict[str(i)] = items
+            i += 1
+
+        self.res_dict = newdict
+        return True
+
+    def find_test_resonators(self, with_VNA=None):
+        """
+        Does a power sweep over the resonators to see if they have a qubit
+        attached or not, and changes the state in the res_dict
+        """
+        if with_VNA is None:
+            try:
+                if self.instr_VNA.get_instr() == '':
+                    with_VNA = False
+                else:
+                    with_VNA = True
+            except:
+                with_VNA = False
+
+        for resonator, items in self.res_dict.items():
+            freq = items[0]
+            state = items[1]
+
+            if state == 'unknown':
+                if with_VNA:
+                    VNA = self.instr_VNA.get_instr()
+                    VNA.start_frequency(freq - 20e6)
+                    VNA.stop_frequency(freq + 20e6)
+
+                self.measure_resonator_power(freqs=np.arange(freq-5e6,
+                                                             freq+5e6, 0.1e6),
+                                             powers=np.arange(-40, 0.1, 10),
+                                             analyze=False)
+                # self.measure_VNA_power_sweep()
+                fit_res = ma.Resonator_Powerscan_Analysis(label='Resonator_power_scan',
+                                                          close_fig=True)
+                shift = fit_res.results[0]
+                freq = fit_res.results[2]
+                power = fit_res.results[1]
+
+                if np.abs(shift) > 100e3:
+                    state = 'qubit_resonator'
+                    self.freq_res(freq)
+                else:
+                    state = 'test_resonator'
+                self.res_dict[resonator][0] = freq
+                self.res_dict[resonator][1] = state
+        return True
+
+    def find_qubit_resonator_fluxline(self, with_VNA=None, verbose=True):
+        if with_VNA is None:
+            try:
+                if self.instr_VNA.get_instr() == '':
+                    with_VNA = False
+                else:
+                    with_VNA = True
+            except:
+                with_VNA = False
+
+        fluxcurrent = self.instr_FluxCtrl.get_instr()
+        for FBL in fluxcurrent.channel_map:
+            fluxcurrent[FBL](0)
+
+        dac_values = np.arange(-20e-3, 20e-3, 2e-3)
+
+        for resonator, items in self.res_dict.items():
+            best_amplitude = 0  # For comparing which one is coupled closest
+            if items[1] == 'qubit_resonator':
+                freq = items[0]
+                if with_VNA:
+                    VNA = self.instr_VNA.get_instr()
+                    VNA.start_frequency(freq-20e6)
+                    VNA.stop_frequency(freq+20e6)
+                freqs = np.arange(freq-5e6, freq+5e6, 0.1e6)
+                for fluxline in fluxcurrent.channel_map:
+                    t_start = time.strftime('%Y%m%d_%H%M%S')
+
+                    self.measure_resonator_frequency_dac_scan(freqs=freqs,
+                                                              dac_values=dac_values,
+                                                              fluxChan=fluxline,
+                                                              analyze=False)
+                    print('Done flux sweep resonator {} ({} GHz) with {}'.format(
+                          resonator, round(freq/1e9, 3), fluxline))
+
+                    ma.TwoD_Analysis(
+                        label='Resonator_dac_scan', normalize=False)
+
+                    timestamp = a_tools.get_timestamps_in_range(t_start,
+                                                                label=self.msmt_suffix)[0]
+
+                    ma.TwoD_Analysis(
+                        label='Resonator_dac_scan', normalize=False)
+
+                    fluxcurrent[fluxline](0)
+
+                    fit_res = ma2.VNA_DAC_Analysis(timestamp)
+                    amplitude = fit_res.dac_fit_res.params['amplitude'].value
+                    sweetspot_current = fit_res.sweet_spot_value
+
+                    items[2][fluxline] = amplitude
+
+                    if amplitude > best_amplitude:
+                        best_amplitude = amplitude
+                        self.cfg_dc_flux_ch(fluxline)
+                        self.res_dict[resonator][3] = 'Q' + fluxline[4]
+                        self.res_dict[resonator][4] = sweetspot_current
+
+                        self.fl_dc_V0(sweetspot_current)
+                        fluxcurrent[fluxline](sweetspot_current)
+
+        if verbose:
+            for items in self.res_dict.values():
+                print('{}, f = {:.3f}, linked to {},'
+                      ' sweetspot current = {:.3f} mA'.format(items[1],
+                                                              items[0]/1e9,
+                                                              items[3],
+                                                              items[4]*1e3))
+        return True
+
+    def find_resonator_sweetspot(self, freqs=None, dac_values=None,
+                                 fluxChan=None, update=True):
+        '''
+        Finds the resonator sweetspot current.
+        TODO: - measure all FBL-resonator combinations
+        TODO: - implement way of distinguishing which fluxline is most coupled
+        TODO: - create method that moves qubits away from sweetspot when they
+                are not being measured (should not move them to some other
+                qubit frequency of course)
+        '''
+        if freqs is None:
+            freq_center = self.freq_res()
+            freq_range = 20e6
+            freqs = np.arange(freq_center-freq_range/2,
+                              freq_center+freq_range/2, 0.5e6)
+
+        if dac_values is None:
+            dac_values = np.linspace(-10e-3, 10e-3, 101)
+
+        if fluxChan is None:
+            if self.cfg_dc_flux_ch() == 1:  # Initial value
+                fluxChan = 'FBL_1'
+            else:
+                fluxChan = self.cfg_dc_flux_ch()
+
+        t_start = time.strftime('%Y%m%d_%H%M%S')
+        self.measure_resonator_frequency_dac_scan(freqs=freqs,
+                                                  dac_values=dac_values,
+                                                  fluxChan=fluxChan,
+                                                  analyze=False)
+        if update:
+
+            import pycqed.analysis_v2.spectroscopy_analysis as sa
+            fit_res = sa.VNA_DAC_Analysis(timestamp=t_start)
+            sweetspot_current = fit_res.sweet_spot_value
+            self.fl_dc_V0(sweetspot_current)
+            fluxcurrent = self.instr_FluxCtrl.get_instr()
+            fluxcurrent[self.cfg_dc_flux_ch()](sweetspot_current)
+
+        return True
+
     def find_resonator_frequency(self, use_min=False,
                                  update=True,
                                  freqs=None,
